test demo of high speed continous ads1115 driver
This demo is to test the continous (instead of single shot r/w every time) high speed ads1115 self-written driver to sample continously  No such lib for ads1115 in c++ found yet, all others are single shot ones.

ps: in this test demo the ads sampling is controlled by Qtimer instead of interrupt. This is not because we can't.  This is to separate the lib and expose if any bug.